requirements.txt
Django==2.1
djangorestframework==3.9.0
psycopg2==2.8
Pillow==6.0.0
celery==4.2.0